feat: layout classes로 구분해서 나누기 #78
