+++ src/apps/room/[id]/index.tsx
@@ -1,17 +1,18 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { useParams, useNavigate } from '@tanstack/react-router';
 import { ArrowLeft } from 'lucide-react';
+import { toast } from 'react-toastify';
 
 import { Button } from '@/shared/components/ui/button';
-import { useResponsive } from '@/shared/hooks/useResponsive';
+import { useResponsiveClasses } from '@/shared/hooks/useResponsive';
 import { accountsQuery, gameRoomQuery } from '@/shared/service/api/query';
 
+import GameInfoCard from './components/GameInfoCard';
 import GameRoomHeader from './components/GameRoomHeader';
-import DesktopLayout from './layouts/DesktopLayout';
-import MobileLayout from './layouts/MobileLayout';
+import PlayersList from './components/PlayersList';
+import ReadyControls from './components/ReadyControls';
 
 export default function GameRoomDetailPage() {
-  const { isDesktop } = useResponsive();
   const { roomId } = useParams({ from: '/room/$roomId' });
   const navigate = useNavigate();
 
@@ -34,11 +35,19 @@ export default function GameRoomDetailPage() {
     navigate({ to: '/lobby', replace: false });
   };
 
-  const layoutProps = {
-    isHost,
-    room,
-    roomId,
+  const canStartGame = room.members.length === room.maxMembersCount;
+
+  const handleStartGame = () => {
+    if (!isHost || !canStartGame) return;
+
+    toast.info('개발중인 기능입니다.');
   };
+
+  const roomLayoutStyles = useResponsiveClasses({
+    mobile: 'flex flex-col ',
+    tablet: 'flex flex-col ',
+    desktop: 'grid grid-cols-3 ',
+  });
 
   if (!currentPlayer) {
     return (
@@ -71,11 +80,31 @@ export default function GameRoomDetailPage() {
           {/* 방 정보 헤더 */}
           <GameRoomHeader room={room} />
 
-          {isDesktop ? (
-            <DesktopLayout {...layoutProps} />
-          ) : (
-            <MobileLayout {...layoutProps} />
-          )}
+          {/* 모바일에서는 세로 배치, 데스크톱에서는 가로 배치 */}
+          <div className={`${roomLayoutStyles} gap-6`}>
+            {/* 왼쪽 영역 - 플레이어 목록 */}
+            <div className="col-span-2">
+              <PlayersList players={room.members} />
+            </div>
+
+            {/* 오른쪽 영역 - 준비 상태 컨트롤과 게임 정보 */}
+            <div className="col-span-1 space-y-4">
+              <ReadyControls
+                isHost={isHost}
+                canStartGame={canStartGame}
+                onStartGame={handleStartGame}
+                currentPlayers={room.members.length}
+                maxPlayers={room.maxMembersCount}
+              />
+
+              {/* 게임 정보 카드 */}
+              <GameInfoCard
+                maxPlayers={room.maxMembersCount}
+                currentPlayers={room.members.length}
+                roomId={roomId}
+              />
+            </div>
+          </div>
         </div>
       </div>
     </div>
